+++ packages/eventcatalog/lib/__tests__/graphs.spec.ts
@@ -25,7 +25,7 @@ describe('graphs', () => {
 
       expect(result.trim()).toEqual(`flowchart LR
 
-My_Event[My Event]:::producer-->My_Service[My Service]:::event
+l-My_Event_2[My Event 2]:::producer-->My_Service[My Service]:::event
 
 classDef event stroke:#2563eb,stroke-width: 4px;
 
@@ -33,11 +33,11 @@ classDef producer stroke:#75d7b6,stroke-width: 2px;
 
 classDef consumer stroke:#818cf8,stroke-width: 2px;
 
-My_Service[My Service]:::event-->My_Event_2[My Event 2]:::consumer
+My_Service[My Service]:::event-->r-My_Event[My Event]:::consumer
 
-click My_Event href "/docs/events/My Event" "Go to My Event" _self
+click l-My_Event_2 href "/docs/events/My Event 2" "Go to My Event 2" _self
 
-click My_Event_2 href "/docs/events/My Event 2" "Go to My Event 2" _self
+click r-My_Event href "/docs/events/My Event" "Go to My Event" _self
 
 click My_Service href "/docs/services/My Service" "Go to My Service" _self`);
     });
@@ -56,7 +56,7 @@ click My_Service href "/docs/services/My Service" "Go to My Service" _self`);
 
       expect(result.trim()).toEqual(`flowchart LR
 
-Service_1[Service 1]:::producer-->My_Event[My Event]:::event
+l-Service_1[Service 1]:::producer-->My_Event[My Event]:::event
 
 classDef event stroke:#2563eb,stroke-width: 4px;
 
@@ -64,11 +64,11 @@ classDef producer stroke:#75d7b6,stroke-width: 2px;
 
 classDef consumer stroke:#818cf8,stroke-width: 2px;
 
-My_Event[My Event]:::event-->Service_2[Service 2]:::consumer
+My_Event[My Event]:::event-->r-Service_2[Service 2]:::consumer
 
-click Service_1 href "/docs/services/Service 1" "Go to Service 1" _self
+click l-Service_1 href "/docs/services/Service 1" "Go to Service 1" _self
 
-click Service_2 href "/docs/services/Service 2" "Go to Service 2" _self
+click r-Service_2 href "/docs/services/Service 2" "Go to Service 2" _self
 
 click My_Event href "/docs/events/My Event" "Go to My Event" _self`);
     });